split CMAv2 by condition type
